[feature] navigation 3 final implementation
