+++ gradle/libs.versions.toml
@@ -1,25 +1,33 @@
 [versions]
 androidGradlePlugin = "8.13.0"
-kotlinVersion = "2.2.20"
-ksp = "2.2.20-2.0.4"
-composeVersion = "1.10.0+dev3015"
-androidx-room = "2.8.2"
+kotlinVersion = "2.2.21"
+ksp = "2.3.0"
+composeVersion = "1.10.0-beta01"
+buildkonfigGradlePlugin = "0.17.1"
+
+androidx-room = "2.8.3"
 sqlite = "2.6.1"
 
 activity-compose = "1.11.0"
-ui-tooling = "1.9.3"
+ui-tooling = "1.9.4"
 appcompat = "1.7.1"
-buildkonfigGradlePlugin = "0.17.1"
+
+material3="1.9.0-beta03"
+material="1.7.3"
+
 coil= "3.3.0"
 core-ktx = "1.17.0"
+
 kotlinxCoroutinesAndroid = "1.10.2"
-ktorVersion = "3.3.1"
+
+ktorVersion = "3.3.2"
+
 koin = "4.1.1"
 
 androidx-navigation = "2.9.1"
-androidx-lifecycle = "2.10.0+dev3015"
-androidx-nav3 = "1.0.0+dev3015"
-androidx-adaptive = "1.2.0+dev3015"
+androidx-lifecycle = "2.10.0-alpha04"
+androidx-nav3 = "1.0.0-alpha04"
+androidx-adaptive = "1.2.0"
 
 kermit = "2.0.8"
 
@@ -30,13 +38,24 @@ androidx-core-ktx = { module = "androidx.core:core-ktx", version.ref = "core-ktx
 androidx-core-animation = { group = "androidx.core", name = "core-animation", version = "1.0.0" }
 androidx-ui-tooling = { group = "androidx.compose.ui", name = "ui-tooling", version.ref = "ui-tooling" }
 
+#Compose Multiplatform
+jetbrains-compose-runtime = { group = "org.jetbrains.compose.runtime", name = "runtime", version.ref = "composeVersion" }
+jetbrains-compose-foundation = { group = "org.jetbrains.compose.foundation", name = "foundation", version.ref = "composeVersion" }
+jetbrains-compose-animation = { group = "org.jetbrains.compose.animation", name = "animation", version.ref = "composeVersion" }
+jetbrains-compose-components-resources = { group = "org.jetbrains.compose.components", name = "components-resources", version.ref = "composeVersion" }
+jetbrains-compose-ui-tooling-preview = { group = "org.jetbrains.compose.ui", name = "ui-tooling-preview", version.ref = "composeVersion" }
+jetbrains-compose-material3 = { group = "org.jetbrains.compose.material3", name = "material3", version.ref = "material3" }
+jetbrains-compose-material-icons-extended = { group = "org.jetbrains.compose.material", name = "material-icons-extended", version.ref = "material" }
+
+
 kermit = { module = "co.touchlab:kermit", version.ref = "kermit" }
+
 kotlinx-coroutines-core = { module = "org.jetbrains.kotlinx:kotlinx-coroutines-core", version.ref = "kotlinxCoroutinesAndroid" }
+kotlinx-coroutines-android = { module = "org.jetbrains.kotlinx:kotlinx-coroutines-android", version.ref = "kotlinxCoroutinesAndroid" }
 
 coil-compose = { module = "io.coil-kt.coil3:coil-compose", version.ref = "coil" }
 coil-network = { module = "io.coil-kt.coil3:coil-network-ktor3", version.ref = "coil" }
 
-kotlinx-coroutines-android = { module = "org.jetbrains.kotlinx:kotlinx-coroutines-android", version.ref = "kotlinxCoroutinesAndroid" }
 ktor-client-content-negotiation = { module = "io.ktor:ktor-client-content-negotiation", version.ref = "ktorVersion" }
 ktor-client-logging = { module = "io.ktor:ktor-client-logging", version.ref = "ktorVersion" }
 ktor-client-core = { module = "io.ktor:ktor-client-core", version.ref = "ktorVersion" }
@@ -49,25 +68,24 @@ koin-android = {module="io.insert-koin:koin-android", version.ref="koin"}
 koin-compose = {module="io.insert-koin:koin-compose", version.ref="koin"}
 koin-composeViewModel = { module = "io.insert-koin:koin-compose-viewmodel", version.ref = "koin" }
 
+androidx-lifecycle-runtime = { module = "org.jetbrains.androidx.lifecycle:lifecycle-runtime-compose", version.ref = "androidx-lifecycle" }
 androidx-lifecycle-viewmodel-compose = { module = "org.jetbrains.androidx.lifecycle:lifecycle-viewmodel-compose", version.ref = "androidx-lifecycle" }
+androidx-lifecycle-viewmodel-nav3 = { module = "org.jetbrains.androidx.lifecycle:lifecycle-viewmodel-navigation3", version.ref = "androidx-lifecycle" }
+
 androidx-navigation-compose = { module = "org.jetbrains.androidx.navigation:navigation-compose", version.ref = "androidx-navigation" }
+
 androidx-nav3-ui = { module = "org.jetbrains.androidx.navigation3:navigation3-ui", version.ref = "androidx-nav3" }
 androidx-material3-adaptive = { module = "org.jetbrains.compose.material3.adaptive:adaptive", version.ref = "androidx-adaptive" }
 
 androidx-room-compiler = { group = "androidx.room", name = "room-compiler", version.ref = "androidx-room" }
 androidx-room-runtime = { group = "androidx.room", name = "room-runtime", version.ref = "androidx-room" }
-androidx-room-ktx = { group = "androidx.room", name = "room-ktx", version.ref = "androidx-room" }
 sqlite-bundled = { module = "androidx.sqlite:sqlite-bundled", version.ref = "sqlite" }
 
 buildkonfig-gradle-plugin = { module = "com.codingfeline.buildkonfig:buildkonfig-gradle-plugin", version.ref = "buildkonfigGradlePlugin" }
 
-[bundles]
-
 [plugins]
 android-application = { id = "com.android.application", version.ref = "androidGradlePlugin" }
 android-library = { id = "com.android.library", version.ref = "androidGradlePlugin" }
-#TODO: this plugin no found for preview
-android-kotlin-multiplatform-library = { id = "com.android.kotlin.multiplatform.library", version.ref = "androidGradlePlugin" }
 kotlin-multiplatform = { id = "org.jetbrains.kotlin.multiplatform", version.ref = "kotlinVersion" }
 kotlin-serialization = { id = "org.jetbrains.kotlin.plugin.serialization", version.ref = "kotlinVersion" }
 compose-compiler = { id = "org.jetbrains.kotlin.plugin.compose", version.ref = "kotlinVersion" }
