+++ shared/src/androidMain/kotlin/com.santimattius.kmp.entertainment/core/ui/themes/Theme.android.kt
@@ -1,0 +1,23 @@
+package com.santimattius.kmp.entertainment.core.ui.themes
+
+import android.app.Activity
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.ui.graphics.toArgb
+import androidx.compose.ui.platform.LocalView
+import androidx.core.view.WindowInsetsControllerCompat
+
+@Composable
+internal actual fun SystemAppearance(
+    isDark: Boolean
+) {
+    val view = LocalView.current
+    LaunchedEffect(isDark) {
+        val window = (view.context as Activity).window
+        WindowInsetsControllerCompat(window, window.decorView).apply {
+            isAppearanceLightStatusBars = isDark
+            isAppearanceLightNavigationBars = isDark
+        }
+    }
+}
