+++ shared/src/commonMain/kotlin/com.santimattius.kmp.entertainment/feature/splash/Splash.kt
@@ -12,14 +12,11 @@ import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.scale
+import androidx.compose.ui.tooling.preview.Preview
 import com.santimattius.kmp.entertainment.shared.generated.resources.Res
 import com.santimattius.kmp.entertainment.shared.generated.resources.watching_a_movie
 import kotlinx.coroutines.delay
 import org.jetbrains.compose.resources.painterResource
-import org.jetbrains.compose.ui.tooling.preview.Preview
-
-private const val SPLASH_IMAGE =
-    "https://img.freepik.com/vector-gratis/produccion-peliculas-cine-composicion-transparente-realista-imagen-aislada-badajo-campos-vacios-ilustracion-vectorial_1284-66163.jpg?w=740&t=st=1694904624~exp=1694905224~hmac=936e49d4984b5c24111a43ff7835c3962d6cdf002d489f5ed196716577f7d2c3"
 
 @Composable
 fun SplashScreen(navigate: () -> Unit) {
