+++ shared/src/commonMain/kotlin/com.santimattius.kmp.entertainment/navigation/NavExtensions.kt
@@ -5,26 +5,14 @@ import androidx.compose.animation.ExperimentalSharedTransitionApi
 import androidx.compose.animation.SharedTransitionLayout
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.CompositionLocalProvider
-import androidx.compose.runtime.MovableContent
 import androidx.navigation.NavBackStackEntry
 import androidx.navigation.NavDeepLink
-import androidx.navigation.NavGraph.Companion.findStartDestination
 import androidx.navigation.NavGraphBuilder
-import androidx.navigation.NavHostController
 import androidx.navigation.compose.composable
 import com.santimattius.kmp.entertainment.core.ui.animation.LocalNavAnimatedVisibilityScope
 import com.santimattius.kmp.entertainment.core.ui.animation.LocalSharedTransitionScope
 
-fun <T : Any> NavHostController.navigatePoppingUpToStartDestination(route: T) {
-    navigate(route) {
-        popUpTo(graph.findStartDestination().navigatorName) {
-            saveState = true
-        }
-        launchSingleTop = true
-        restoreState = true
-    }
-}
-
+//TODO: test this with nav3
 inline fun <reified T : Any> NavGraphBuilder.composableNavAnimated(
     deepLinks: List<NavDeepLink> = emptyList(),
     noinline content: @Composable AnimatedContentScope.(NavBackStackEntry) -> Unit
@@ -40,6 +28,7 @@ inline fun <reified T : Any> NavGraphBuilder.composableNavAnimated(
     }
 }
 
+//TODO: test this with nav3
 @OptIn(ExperimentalSharedTransitionApi::class)
 @Composable
 fun AnimatedTransactionLayout(content: @Composable () -> Unit) {
