+++ shared/src/iosMain/kotlin/com.santimattius.kmp.entertainment/core/ui/themes/Theme.ios.kt
@@ -1,0 +1,17 @@
+package com.santimattius.kmp.entertainment.core.ui.themes
+
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
+import platform.UIKit.UIApplication
+import platform.UIKit.UIStatusBarStyleDarkContent
+import platform.UIKit.UIStatusBarStyleLightContent
+import platform.UIKit.setStatusBarStyle
+
+@Composable
+internal actual fun SystemAppearance(isDark: Boolean) {
+    LaunchedEffect(isDark) {
+        UIApplication.sharedApplication.setStatusBarStyle(
+            if (isDark) UIStatusBarStyleDarkContent else UIStatusBarStyleLightContent
+        )
+    }
+}
